pyproject.toml
[project]
name = "excalia-autovote"
version = "0.1.0"
description = ""
authors = [
    {name = "WallansDev",email = "timothe.vaquie1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "python-dotenv>=1.0.0",
]

[tool.poetry]
packages = [{include = "excalia_autovote", from = "src"}]

[project.scripts]
excalia-autovote = "excalia_autovote.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
